Move check-in data get to custom hook
